Fix forms, moved validate_eth_address to utils
